#43 コードが汚く見づらかったので投稿ボタンと更新ボタンからロジックの切り出しをした。
