--- src/pages/Register/RegisterTatoeChild/RegisterTatoeCreateBtn.tsx
@@ -1,150 +1,0 @@
-import { useRouter } from 'next/router';
-import { useRecoilState, useRecoilValue } from 'recoil';
-import { TatoeAtom } from '../../../components/utils/atoms/TatoeAtom';
-import { useUserInfo } from '../../../components/hooks/useUserInfo';
-import { useAuth } from '../../../components/hooks/useAuth';
-import { Tatoe } from '../../../components/types/types';
-import { LoginUserAtom } from '../../../components/utils/atoms/LoginUserAtom';
-
-export const RegisterTatoeCreateBtn = (props: Tatoe) => {
-  const {
-    query_tId,
-    title,
-    shortParaphrase,
-    description,
-    createdAt,
-    updatedAt,
-  } = props;
-  const { userId } = useAuth();
-  const { user } = useUserInfo(userId);
-
-  const [tatoe, setTatoe] = useRecoilState(TatoeAtom);
-
-  const persistAccessToken = useRecoilValue(LoginUserAtom);
-
-  const router = useRouter();
-
-  // TODO ユーザーが title, description, shortParaphrase 入力したら async await して API にデータを送る
-  const submitTatoe = async (): Promise<string> => {
-    if (!userId || !user) {
-      return null;
-    }
-    if (title === '' || shortParaphrase === '' || description === '') {
-      alert('入力されていない箇所があります。');
-      return;
-    }
-
-    if (user.userName === '') {
-      alert('ユーザー名を登録して投稿してください。');
-      return;
-    }
-
-    // // 初回登録
-    if (!query_tId) {
-      const res = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/tatoe`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${persistAccessToken}`,
-        },
-        body: JSON.stringify({
-          title,
-          shortParaphrase,
-          description,
-        }),
-      });
-      const { data } = await res.json();
-      // フロント側で使えるようにデータ加工
-      const formattedData: Tatoe = {
-        tId: data.id,
-        userId: data.userId,
-        createdAt: data.createdAt,
-        updatedAt: data.updatedAt,
-        title: data.title,
-        description: data.description,
-        shortParaphrase: data.shortParaphrase,
-      };
-      const newTatoe = [formattedData, ...tatoe];
-
-      setTatoe(newTatoe);
-
-      router.push({
-        pathname: '/DashBoard/UserTatoeList',
-      });
-    }
-    // 更新時
-    if (query_tId) {
-      tatoe.map(async (item: Tatoe) => {
-        if (item.tId === query_tId) {
-          const tId = item.tId;
-          const res = await fetch(
-            `${process.env.NEXT_PUBLIC_BACKEND_URL}/tatoe/${tId}`,
-            {
-              method: 'PUT',
-              headers: {
-                'Content-Type': 'application/json',
-                Authorization: `Bearer ${persistAccessToken}`,
-              },
-              body: JSON.stringify({
-                title,
-                shortParaphrase,
-                description,
-                tId,
-                createdAt,
-                updatedAt,
-              }),
-            }
-          );
-          const { data } = await res.json();
-
-          const updatedTatoe = tatoe.map((item: Tatoe) => {
-            if (item.tId === query_tId) {
-              return {
-                tId: item.tId,
-                userId: data.userId,
-                createdAt: data.createdAt,
-                updatedAt: data.updatedAt,
-                title: data.title,
-                description: data.description,
-                shortParaphrase: data.shortParaphrase,
-              };
-            }
-            return item;
-          });
-
-          setTatoe(updatedTatoe);
-        }
-      });
-
-      tatoe.map((item: Tatoe) => {
-        if (item.tId === query_tId) {
-          router.push({
-            pathname: '/DashBoard/UserTatoeList',
-          });
-        }
-      });
-    }
-  };
-
-  return (
-    <div className='flex justify-end'>
-      <button
-        onClick={submitTatoe}
-        // onClick={props.onClick}
-        // type='submit'
-        className='
-      p-3
-      w-[200px]
-      rounded-full
-      bg-dark_green
-      text-gray-800
-      text-lg
-      hover:bg-opacity-90
-    '
-      >
-        投稿する
-        {/* {props.children} */}
-      </button>
-    </div>
-  );
-};
